mysite/requirements.txt
flask
requests
pyyaml
pyautogui
pillow
pyjwt
